# docs/openapi.yml
openapi: 3.1.0
info:
  title: Allama API
  summary: Allama API
  description: Allama is the open source automation platform for enterprise. You can
    operate Allama in headless mode by using the API to create, manage, and run
    workflows.
  termsOfService: https://docs.google.com/document/d/e/2PACX-1vQvDe3SoVAPoQc51MgfGCP71IqFYX_rMVEde8zC4qmBCec5f8PLKQRdxa6tsUABT8gWAR9J-EVs2CrQ/pub
  contact:
    name: Allama Founders
    email: founders@allama.com
  license:
    name: AGPL-3.0
    url: https://www.gnu.org/licenses/agpl-3.0.html
  version: '1'
paths:
...
  /secrets/search:
    get:
      tags:
      - secrets
      summary: Search Secrets
      description: Search secrets.
      operationId: secrets-search_secrets
      security:
      - APIKeyCookie: []
      - APIKeyHeader: []
      parameters:
      - name: environment
        in: query
        required: true
        schema:
          type: string
          title: Environment
      - name: name
        in: query
        required: false
        schema:
          anyOf:
          - type: array
            uniqueItems: true
            items:
              type: string
          - type: 'null'
          description: Filter by secret name
          title: Name
        description: Filter by secret name
      - name: id
        in: query
        required: false
        schema:
          anyOf:
          - type: array
            uniqueItems: true
            items:
              type: string
              pattern: secret-[0-9a-f]{32}
          - type: 'null'
          description: Filter by secret ID
          title: Id
        description: Filter by secret ID
      - name: type
        in: query
        required: false
        schema:
          anyOf:
          - type: array
            uniqueItems: true
            items:
              $ref: '#/components/schemas/SecretType'
          - type: 'null'
          description: Filter by secret type
          title: Type
        description: Filter by secret type
      - name: workspace_id
        in: query
        required: true
        schema:
          type: string
          format: uuid
          title: Workspace Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/SecretRead'
                title: Response Secrets-Search Secrets
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /secrets:
    get:
      tags:
      - secrets
      summary: List Secrets
      description: List user secrets.
      operationId: secrets-list_secrets
      security:
      - APIKeyCookie: []
      - APIKeyHeader: []
      parameters:
      - name: type
        in: query
        required: false
        schema:
          anyOf:
          - type: array
            uniqueItems: true
            items:
              $ref: '#/components/schemas/SecretType'
          - type: 'null'
          description: Filter by secret type
          title: Type
        description: Filter by secret type
      - name: workspace_id
        in: query
        required: true
        schema:
          type: string
          format: uuid
          title: Workspace Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/SecretReadMinimal'
                title: Response Secrets-List Secrets
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    post:
      tags:
      - secrets
      summary: Create Secret
      description: Create a secret.
      operationId: secrets-create_secret
      security:
      - APIKeyCookie: []
      - APIKeyHeader: []
      parameters:
      - name: workspace_id
        in: query
        required: true
        schema:
          type: string
          format: uuid
          title: Workspace Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SecretCreate'
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /secrets/{secret_name}:
    get:
      tags:
      - secrets
      summary: Get Secret By Name
      description: Get a secret.
      operationId: secrets-get_secret_by_name
      security:
      - APIKeyCookie: []
      - APIKeyHeader: []
      parameters:
      - name: secret_name
        in: path
        required: true
        schema:
          type: string
          title: Secret Name
      - name: workspace_id
        in: query
        required: true
        schema:
          type: string
          format: uuid
          title: Workspace Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SecretRead'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /secrets/{secret_id}:
    post:
      tags:
      - secrets
      summary: Update Secret By Id
      description: Update a secret by ID.
      operationId: secrets-update_secret_by_id
      security:
      - APIKeyCookie: []
      - APIKeyHeader: []
      parameters:
      - name: secret_id
        in: path
        required: true
        schema:
          type: string
          pattern: secret-[0-9a-f]{32}
          title: Secret Id
      - name: workspace_id
        in: query
        required: true
        schema:
          type: string
          format: uuid
          title: Workspace Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SecretUpdate'
      responses:
        '204':
          description: Successful Response
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    delete:
      tags:
      - secrets
      summary: Delete Secret By Id
      description: Delete a secret by ID.
      operationId: secrets-delete_secret_by_id
      security:
      - APIKeyCookie: []
      - APIKeyHeader: []
      parameters:
      - name: secret_id
        in: path
        required: true
        schema:
          type: string
          pattern: secret-[0-9a-f]{32}
          title: Secret Id
      - name: workspace_id
        in: query
        required: true
        schema:
          type: string
          format: uuid
          title: Workspace Id
      responses:
        '204':
          description: Successful Response
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
  /organization/secrets:
    get:
      tags:
      - organization-secrets
      summary: List Org Secrets
      description: List organization secrets.
      operationId: organization-secrets-list_org_secrets
      security:
      - APIKeyCookie: []
      - APIKeyHeader: []
      parameters:
      - name: type
        in: query
        required: false
        schema:
          anyOf:
          - type: array
            uniqueItems: true
            items:
              $ref: '#/components/schemas/SecretType'
          - type: 'null'
          description: Filter by secret type
          title: Type
        description: Filter by secret type
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/SecretReadMinimal'
                title: Response Organization-Secrets-List Org Secrets
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    post:
      tags:
      - organization-secrets
      summary: Create Org Secret
      description: Create an organization secret.
      operationId: organization-secrets-create_org_secret
      security:
      - APIKeyCookie: []
      - APIKeyHeader: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SecretCreate'
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /organization/secrets/{secret_name}:
    get:
      tags:
      - organization-secrets
      summary: Get Org Secret By Name
      description: Get an organization secret by name.
      operationId: organization-secrets-get_org_secret_by_name
      security:
      - APIKeyCookie: []
      - APIKeyHeader: []
      parameters:
      - name: secret_name
        in: path
        required: true
        schema:
          type: string
          title: Secret Name
      - name: environment
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Environment
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SecretRead'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /organization/secrets/{secret_id}:
    post:
      tags:
      - organization-secrets
      summary: Update Org Secret By Id
      description: Update an organization secret by ID.
      operationId: organization-secrets-update_org_secret_by_id
      security:
      - APIKeyCookie: []
      - APIKeyHeader: []
      parameters:
      - name: secret_id
        in: path
        required: true
        schema:
          type: string
          pattern: secret-[0-9a-f]{32}
          title: Secret Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SecretUpdate'
      responses:
        '204':
          description: Successful Response
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    delete:
      tags:
      - organization-secrets
      summary: Delete Org Secret By Id
      description: Delete an organization secret by ID.
      operationId: organization-secrets-delete_org_secret_by_id
      security:
      - APIKeyCookie: []
      - APIKeyHeader: []
      parameters:
      - name: secret_id
        in: path
        required: true
        schema:
          type: string
          pattern: secret-[0-9a-f]{32}
          title: Secret Id
      responses:
        '204':
          description: Successful Response
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
components:
  schemas:
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    SecretCreate:
      properties:
        type:
          $ref: '#/components/schemas/SecretType'
          default: custom
        name:
          type: string
          maxLength: 100
          minLength: 1
          title: Name
        description:
          anyOf:
          - type: string
            maxLength: 1000
            minLength: 0
          - type: 'null'
          title: Description
        keys:
          items:
            $ref: '#/components/schemas/SecretKeyValue'
          type: array
          maxItems: 100
          minItems: 1
          title: Keys
        tags:
          anyOf:
          - additionalProperties:
              type: string
            type: object
          - type: 'null'
          title: Tags
        environment:
          type: string
          title: Environment
          default: default
      type: object
      required:
      - name
      - keys
      title: SecretCreate
      description: 'Create a new secret.


        Secret types

        ------------

        - `custom`: Arbitrary user-defined types

        - `token`: A token, e.g. API Key, JWT Token (TBC)

        - `oauth2`: OAuth2 Client Credentials (TBC)'
    SecretKeyValue:
      properties:
        key:
          type: string
          title: Key
        value:
          type: string
          format: password
          title: Value
          writeOnly: true
      type: object
      required:
      - key
      - value
      title: SecretKeyValue
    SecretRead:
      properties:
        id:
          type: string
          title: Id
        type:
          $ref: '#/components/schemas/SecretType'
        name:
          type: string
          title: Name
        description:
          anyOf:
          - type: string
          - type: 'null'
          title: Description
        encrypted_keys:
          type: string
          format: binary
          title: Encrypted Keys
        environment:
          type: string
          title: Environment
        tags:
          anyOf:
          - additionalProperties:
              type: string
            type: object
          - type: 'null'
          title: Tags
        owner_id:
          type: string
          format: uuid
          title: Owner Id
        created_at:
          type: string
          format: date-time
          title: Created At
        updated_at:
          type: string
          format: date-time
          title: Updated At
      type: object
      required:
      - id
      - type
      - name
      - encrypted_keys
      - environment
      - owner_id
      - created_at
      - updated_at
      title: SecretRead
    SecretReadMinimal:
      properties:
        id:
          type: string
          title: Id
        type:
          $ref: '#/components/schemas/SecretType'
        name:
          type: string
          title: Name
        description:
          anyOf:
          - type: string
          - type: 'null'
          title: Description
        keys:
          items:
            type: string
          type: array
          title: Keys
        environment:
          type: string
          title: Environment
      type: object
      required:
      - id
      - type
      - name
      - keys
      - environment
      title: SecretReadMinimal
    SecretType:
      type: string
      enum:
      - custom
      - ssh-key
      - mtls
      - ca-cert
      title: SecretType
      description: The type of a secret.
    SecretUpdate:
      properties:
        type:
          anyOf:
          - $ref: '#/components/schemas/SecretType'
          - type: 'null'
        name:
          anyOf:
          - type: string
            maxLength: 100
            minLength: 1
          - type: 'null'
          title: Name
        description:
          anyOf:
          - type: string
            maxLength: 1000
            minLength: 0
          - type: 'null'
          title: Description
        keys:
          anyOf:
          - items:
              $ref: '#/components/schemas/SecretKeyValue'
            type: array
            maxItems: 100
            minItems: 1
          - type: 'null'
          title: Keys
        tags:
          anyOf:
          - additionalProperties:
              type: string
            type: object
            maxProperties: 1000
            minProperties: 0
          - type: 'null'
          title: Tags
        environment:
          anyOf:
          - type: string
            maxLength: 100
            minLength: 1
          - type: 'null'
          title: Environment
      type: object
      title: SecretUpdate
      description: 'Update a secret.


        Secret types

        ------------

        - `custom`: Arbitrary user-defined types

        - `token`: A token, e.g. API Key, JWT Token (TBC)

        - `oauth2`: OAuth2 Client Credentials (TBC)'
...
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError
...
  securitySchemes:
    APIKeyCookie:
      type: apiKey
      in: cookie
      name: fastapiusersauth
    APIKeyHeader:
      type: apiKey
      in: header
      name: x-allama-service-key
...